+++ src/main/java/org/stellar/sdk/operations/LiquidityPoolDepositOperation.java
@@ -8,8 +8,9 @@ import lombok.ToString;
 import lombok.experimental.SuperBuilder;
 import org.stellar.sdk.AssetAmount;
 import org.stellar.sdk.LiquidityPool;
-import org.stellar.sdk.LiquidityPoolID;
 import org.stellar.sdk.Price;
+import org.stellar.sdk.Util;
+import org.stellar.sdk.xdr.Hash;
 import org.stellar.sdk.xdr.LiquidityPoolDepositOp;
 import org.stellar.sdk.xdr.Operation.OperationBody;
 import org.stellar.sdk.xdr.OperationType;
@@ -26,7 +27,7 @@ import org.stellar.sdk.xdr.OperationType;
 @SuperBuilder(toBuilder = true)
 public class LiquidityPoolDepositOperation extends Operation {
   /** The liquidity pool ID. * */
-  @NonNull private final LiquidityPoolID liquidityPoolID;
+  @NonNull private final String liquidityPoolID;
 
   /** Maximum amount of first asset to deposit. * */
   @NonNull private final String maxAmountA;
@@ -63,7 +64,8 @@ public class LiquidityPoolDepositOperation extends Operation {
    * @return {@link LiquidityPoolDepositOperation} object
    */
   public static LiquidityPoolDepositOperation fromXdr(LiquidityPoolDepositOp op) {
-    LiquidityPoolID liquidityPoolID = LiquidityPoolID.fromXdr(op.getLiquidityPoolID());
+    String liquidityPoolID =
+        Util.bytesToHex(op.getLiquidityPoolID().getPoolID().getHash()).toLowerCase();
     String maxAmountA = Operation.fromXdrAmount(op.getMaxAmountA().getInt64());
     String maxAmountB = Operation.fromXdrAmount(op.getMaxAmountB().getInt64());
     Price minPrice = Price.fromXdr(op.getMinPrice());
@@ -75,7 +77,8 @@ public class LiquidityPoolDepositOperation extends Operation {
   @Override
   OperationBody toOperationBody() {
     LiquidityPoolDepositOp op = new LiquidityPoolDepositOp();
-    op.setLiquidityPoolID(this.getLiquidityPoolID().toXdr());
+    op.setLiquidityPoolID(
+        new org.stellar.sdk.xdr.PoolID(new Hash(Util.hexToBytes(this.getLiquidityPoolID()))));
     op.setMaxAmountA(new org.stellar.sdk.xdr.Int64(Operation.toXdrAmount(this.getMaxAmountA())));
     op.setMaxAmountB(new org.stellar.sdk.xdr.Int64(Operation.toXdrAmount(this.getMaxAmountB())));
     op.setMinPrice(this.getMinPrice().toXdr());
@@ -85,5 +88,17 @@ public class LiquidityPoolDepositOperation extends Operation {
     body.setDiscriminant(OperationType.LIQUIDITY_POOL_DEPOSIT);
     body.setLiquidityPoolDepositOp(op);
     return body;
+  }
+
+  // TODO: add test
+  // TODO: equal? toXdr?
+  public abstract static class LiquidityPoolDepositOperationBuilder<
+          C extends LiquidityPoolDepositOperation,
+          B extends LiquidityPoolDepositOperation.LiquidityPoolDepositOperationBuilder<C, B>>
+      extends OperationBuilder<C, B> {
+    public B liquidityPoolID(@NonNull String liquidityPoolID) {
+      this.liquidityPoolID = liquidityPoolID.toLowerCase();
+      return self();
+    }
   }
 }