+++ src/test/java/org/stellar/sdk/StrKeyTest.java
@@ -4,7 +4,6 @@ import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
-import java.io.IOException;
 import org.junit.Test;
 import org.stellar.sdk.exception.FormatException;
 import org.stellar.sdk.xdr.AccountID;
@@ -13,7 +12,7 @@ import org.stellar.sdk.xdr.MuxedAccount;
 
 public class StrKeyTest {
   @Test
-  public void testDecodeEncode() throws IOException, FormatException {
+  public void testDecodeEncode() throws FormatException {
 
     String seed = "SDJHRQF4GCMIIKAAAQ6IHY42X73FQFLHUULAPSKKD4DFDM7UXWWCRHBE";
     byte[] secret = StrKey.decodeCheck(StrKey.VersionByte.SEED, seed.toCharArray());