+++ src/test/java/org/stellar/sdk/requests/TradesRequestBuilderTest.java
@@ -5,7 +5,6 @@ import static org.stellar.sdk.Asset.createNonNativeAsset;
 
 import okhttp3.HttpUrl;
 import org.junit.Test;
-import org.stellar.sdk.LiquidityPoolID;
 import org.stellar.sdk.Server;
 
 public class TradesRequestBuilderTest {
@@ -70,9 +69,7 @@ public class TradesRequestBuilderTest {
     HttpUrl uri =
         server
             .trades()
-            .forLiquidityPool(
-                new LiquidityPoolID(
-                    "67260c4c1807b262ff851b0a3fe141194936bb0215b2f77447f1df11998eabb9"))
+            .forLiquidityPool("67260c4c1807b262ff851b0a3fe141194936bb0215b2f77447f1df11998eabb9")
             .buildUri();
     assertEquals(
         "https://horizon-testnet.stellar.org/liquidity_pools/67260c4c1807b262ff851b0a3fe141194936bb0215b2f77447f1df11998eabb9/trades",